requirements_packages.txt
flask>=2.0.0
torch>=1.8.0
transformers>=4.0.0
pillow>=8.0.0
werkzeug>=2.0.0